refactor: extract useVariables custom hook to reuse the logic in the sendRequest component
